added gitignore for visual studio projects and c++
updated project so it builds.  by no means does it do anything as I am still writing the pseudo code examples.  Some of it doesn't make sense in the recursive call.

the += of actions which are linked list need to managed correctly on return.
memory management needs to be sorted out for UpdateResult.
updatedown passes a parameter state, however the "this" command is in the hsm which doesn't inherit state.  I am guessing this is meant to say m_pCurrentState or the parameter should be HSM.  I am guessing the latter is true given the method's class.